jets: separates +scot from +scow, enabling the old tape jets
