+++ pkg/urbit/jets/tree.c
@@ -1059,11 +1059,8 @@ static u3j_core _140_qua_d[] =
   { "mole", 7, _140_qua_mole_a, 0, _140_qua_mole_ha },
   { "mule", 7, _140_qua_mule_a, 0, _140_qua_mule_ha },
 
-  //  XX disabled, implicated in memory corruption
-  //  write tests and re-enable
-  //
-  // { "scot", 7, _140_qua_scot_a, 0, _140_qua_scot_ha },
-  // { "scow", 7, _140_qua_scow_a, 0, _140_qua_scow_ha },
+  { "scot", 7, _140_qua_scot_a, 0, _140_qua_scot_ha },
+  { "scow", 7, _140_qua_scow_a, 0, _140_qua_scow_ha },
   { "slaw", 7, _140_qua_slaw_a, 0, _140_qua_slaw_ha },
   {}
 };
